feat: add config support for list,activate,deploy (#15)
This commit adds config file support for the list, activate and deploy
command.

It also deprecates the --functions-env flag

BREAKING CHANGE: Deprecating --functions-env as an option

fix: #15, fix: #27
