+++ src/commands/list.ts
@@ -1,120 +1,16 @@
-import {
-  ListConfig as ApiListConfig,
-  ListOptions,
-  TwilioServerlessApiClient,
-} from '@twilio-labs/serverless-api';
+import { TwilioServerlessApiClient } from '@twilio-labs/serverless-api';
 import chalk from 'chalk';
 import debug from 'debug';
-import dotenv from 'dotenv';
-import path from 'path';
-import { PackageJson } from 'type-fest';
-import { Arguments, Argv } from 'yargs';
+import { Argv } from 'yargs';
 import { checkConfigForCredentials } from '../checks/check-credentials';
 import checkForValidServiceSid from '../checks/check-service-sid';
+import { getConfigFromFlags, ListCliFlags, ListConfig } from '../config/list';
 import { printListResult } from '../printers/list';
-import { getFunctionServiceSid } from '../serverless-api/utils';
-import { fileExists, readFile } from '../utils/fs';
-import { sharedCliOptions, SharedFlags } from './shared';
+import { sharedCliOptions } from './shared';
 import { CliInfo } from './types';
-import { deprecateProjectName, getFullCommand } from './utils';
+import { getFullCommand } from './utils';
 
 const log = debug('twilio-run:list');
-
-export type ListConfig = ApiListConfig & {
-  cwd: string;
-  properties?: string[];
-  extendedOutput: boolean;
-};
-
-export type ListCliFlags = Arguments<
-  SharedFlags & {
-    types: string;
-    projectName?: string;
-    serviceName?: string;
-    properties?: string;
-    extendedOutput: boolean;
-    cwd?: string;
-    environment?: string;
-    accountSid?: string;
-    authToken?: string;
-    serviceSid?: string;
-    env?: string;
-  }
-> & {
-  _cliDefault?: {
-    username: string;
-    password: string;
-  };
-};
-
-async function getConfigFromFlags(flags: ListCliFlags): Promise<ListConfig> {
-  const cwd = flags.cwd ? path.resolve(flags.cwd) : process.cwd();
-
-  let { accountSid, authToken } = flags;
-  if (!accountSid || !authToken) {
-    const envPath = path.resolve(cwd, flags.env || '.env');
-
-    let contentEnvFile;
-    if (!(await fileExists(envPath))) {
-      contentEnvFile = '';
-    } else {
-      contentEnvFile = await readFile(envPath, 'utf8');
-    }
-
-    const localEnv = dotenv.parse(contentEnvFile);
-
-    accountSid =
-      flags.accountSid ||
-      localEnv.ACCOUNT_SID ||
-      (flags._cliDefault && flags._cliDefault.username) ||
-      '';
-    authToken =
-      flags.authToken ||
-      localEnv.AUTH_TOKEN ||
-      (flags._cliDefault && flags._cliDefault.password) ||
-      '';
-  }
-
-  const serviceSid =
-    flags.serviceSid ||
-    (await getFunctionServiceSid(cwd, flags.config, 'listConfig'));
-
-  let serviceName = flags.serviceName;
-
-  if (typeof flags.projectName !== 'undefined') {
-    deprecateProjectName();
-    if (!serviceName) {
-      serviceName = flags.projectName;
-    }
-  }
-
-  if (!serviceName) {
-    const pkgJsonPath = path.join(cwd, 'package.json');
-    if (await fileExists(pkgJsonPath)) {
-      const pkgContent = await readFile(pkgJsonPath, 'utf8');
-      const pkgJson: PackageJson = JSON.parse(pkgContent);
-      if (typeof pkgJson.name === 'string') {
-        serviceName = pkgJson.name;
-      }
-    }
-  }
-
-  const types = flags.types.split(',') as ListOptions[];
-
-  return {
-    cwd,
-    accountSid,
-    authToken,
-    serviceSid,
-    serviceName,
-    environment: flags.environment,
-    properties: flags.properties
-      ? flags.properties.split(',').map(x => x.trim())
-      : undefined,
-    extendedOutput: flags.extendedOutput,
-    types,
-  };
-}
 
 function logError(msg: string) {
   console.error(chalk`{red.bold ERROR} ${msg}`);
@@ -127,14 +23,6 @@ function handleError(err: Error) {
 }
 
 export async function handler(flags: ListCliFlags): Promise<void> {
-  // console.log('hi');
-  // const c = readSpecializedConfig(process.cwd(), 'listConfig', {
-  //   projectId: 'ACc2bdaa19578061b45a518a9dedb5e406',
-  //   // environmentSuffix: 'dev',
-  // });
-  // console.log(c);
-  // process.exit(0);
-
   let config: ListConfig;
   try {
     config = await getConfigFromFlags(flags);
